Support cloning to a multicast group in simple_switch
The PSA architecture has recently introduced the ability to clone a
packet (I2E & E2E) to a set of egress ports. P4Runtime also supports
this by offering a runtime API to map a clone session id to a list of
(port, rid) tuples.

While neither PSA nor P4Runtime require this functionality to be
implemented using multicast (by associating a clone session id to a
multicast group), it seems that a lot of HW targets would probably
choose this approach. As a result, this is also the approach we take for
simple_switch. Note that simple_switch now supports cloning to both a
unicast port & a multicast group simultaneously, which may not be the
case for a lot of targets.

The Thrift service for simple_switch has been updated to include new
RPCs for cloning / mirroring. With these new RPCs, it is possible to map
a session id to a unicast port, a multicast group or both. The old RPCs
have been preserved for backward-compatibility but they should be
considered deprecated. The simple_switch_CLI has also been updated; it
uses separate commands to configure a session with unicast and with
multicast (there is currently no way to configure a session for both
using the CLI, but this may be added in the future if needed).